Fix request mock
